[B] refactor: use abstract class
